Add partitioned schemas for native postgres
Alter table commands on partitioned tables cannot be ONLY as they need to
be propagated to the partitions themselves.
In partitioned tables the columns covered by the primary key and unique constraints must include
the partition key columns.
Some foreign keys are not possible (in comparison to the psql_native
schema), because some column-sets cannot longer be defined as unique
without the partition key.